% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\docType{data}
\name{corpus}
\alias{corpus}
\title{English rendering of Indian Independence Day speeches}
\format{
A data frame with 74 rows and 8 variables:
\describe{
\item{year}{year of speech as an integer}
\item{pm}{name of Prime Minister delivering speech}
\item{party}{party of Prime Minister delivering speech}
\item{title}{earlier speeches often were given a title}
\item{footnote}{earlier speeches often have a footnote; newer speeches convey this same information as the title}
\item{source}{where the speech text was found}
\item{url}{where applicable}
\item{text}{full text of English rendering}
}
}
\source{
{see the package README for details on data collection}
}
\usage{
corpus
}
\description{
A dataset containing the full text of Indian Independence Day speeches
}
\keyword{datasets}
